% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacterizeClusters.R
\name{characterizeClusters}
\alias{characterizeClusters}
\title{Title}
\usage{
characterizeClusters(outputFolder, labels)
}
\arguments{
\item{outputFolder}{}

\item{labels}{}
}
\value{

}
\description{
Title
}
