% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{clusterPatients}
\alias{clusterPatients}
\title{Title}
\usage{
clusterPatients(outputFolder, algorithm = "umap-dbscan")
}
\arguments{
\item{outputFolder}{}

\item{algorithm}{}
}
\value{

}
\description{
Title
}
