% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeDistance.R
\name{computeDistance}
\alias{computeDistance}
\title{Title}
\usage{
computeDistance(
  outputFolder,
  minFeaturesPerPerson = 100,
  minPersonsPerFeature = 10
)
}
\arguments{
\item{outputFolder}{}

\item{minFeaturesPerPerson}{}

\item{minPersonsPerFeature}{}
}
\value{

}
\description{
Title
}
