% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{labelByCovariate}
\alias{labelByCovariate}
\title{Title}
\usage{
labelByCovariate(outputFolder, conceptIds = c(201254, 201826))
}
\arguments{
\item{outputFolder}{}

\item{conceptIds}{}
}
\value{

}
\description{
Title
}
